+++ src/main/java/Servlet/SalesInvoiceServlet.java
@@ -1,10 +1,14 @@
 package Servlet;
 
 
+import DTO.InsufficientStockDTO;
 import DTO.SalesInvoiceDTO;
+import Model.InvoiceLineItem;
 import Model.SalesInvoice;
 import Service.SalesInvoiceServiceImpl;
+import Util.PrefixValidator;
 import com.google.gson.Gson;
+import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 
 import javax.servlet.annotation.WebServlet;
@@ -15,7 +19,9 @@ import java.io.IOException;
 import java.io.PrintWriter;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 
 @WebServlet("/salesInvoice/*")
 public class SalesInvoiceServlet extends HttpServlet {
@@ -64,6 +70,12 @@ public class SalesInvoiceServlet extends HttpServlet {
     private void handleGetInvoiceById(String idStr, HttpServletResponse response) throws IOException {
         JsonObject json = new JsonObject();
         try {
+            PrefixValidator validator = new PrefixValidator();
+            String error = validator.validatePrefixedId(idStr , PrefixValidator.EntityType.INVOICE);
+            if (error != null){
+                throw new NumberFormatException(error);
+            }
+
             Integer id = Integer.parseInt(idStr.substring(4));
             SalesInvoice unMaskedinvoice = salesInvoiceService.getInvoiceById(id);
             SalesInvoiceDTO maskedInvoice = new SalesInvoiceDTO(unMaskedinvoice);
@@ -107,12 +119,45 @@ public class SalesInvoiceServlet extends HttpServlet {
         try {
             SalesInvoiceDTO dto = gson.fromJson(request.getReader(), SalesInvoiceDTO.class);
             System.out.println(dto.toString());
-//            for(InvoiceLineItemDTO bdto : dto.getInvoiceLineItems()) System.out.println(bdto.toString());
-//            System.out.println();
+
+            Map<String, List<String>> errors = salesInvoiceService.validate(dto, false);
+            if (!errors.isEmpty()) {
+                response.setStatus(HttpServletResponse.SC_BAD_REQUEST); // 400
+                gson.toJson(Collections.singletonMap("errors", errors), response.getWriter());
+                return;
+            }
+
             SalesInvoice salesInvoiceUnMasked = new SalesInvoice(dto);
             System.out.println(salesInvoiceUnMasked.toString());
-//            for(InvoiceLineItem bdto : salesInvoiceUnMasked.getInvoice_line_items()) System.out.println(bdto.toString());
-//            System.out.println();
+
+            List<InvoiceLineItem> duplicates = salesInvoiceService.findDuplicateLineItems(salesInvoiceUnMasked.getInvoice_line_items());
+            if (!duplicates.isEmpty()) {
+                json.addProperty("status", "error");
+                json.addProperty("message", "Duplicate line items found");
+
+                JsonArray duplicateArray = new JsonArray();
+                for (InvoiceLineItem dup : duplicates)
+                    duplicateArray.add(gson.toJsonTree(dup));
+
+                json.add("duplicates", duplicateArray);
+                response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+                writeResponse(response, json);
+                return; // Stop further processing
+            }
+
+            // Validate stock availability BEFORE creating invoice
+            if(salesInvoiceUnMasked.getStatus() == 1){
+                List<InsufficientStockDTO> insufficientList = salesInvoiceService.validateStockAvailability(salesInvoiceUnMasked.getInvoice_line_items());
+
+                if (!insufficientList.isEmpty()) {
+                    response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+                    json.addProperty("status", "error");
+                    json.addProperty("message", "Insufficient stock for one or more products.");
+                    json.add("insufficient_stock", gson.toJsonTree(insufficientList));
+                    writeResponse(response, json);
+                    return;
+                }
+            }
 
             SalesInvoice insertedInvoice = salesInvoiceService.createInvoice(salesInvoiceUnMasked);
             System.out.println(insertedInvoice.toString());
@@ -158,9 +203,58 @@ public class SalesInvoiceServlet extends HttpServlet {
         }
         String idstr = pathInfo.substring(1);
         try {
+            PrefixValidator validator = new PrefixValidator();
+            String error = validator.validatePrefixedId(idstr , PrefixValidator.EntityType.INVOICE);
+            if (error != null){
+                throw new NumberFormatException(error);
+            }
+            if (salesInvoiceService.isDelivered(Integer.parseInt(idstr))) {
+                response.setStatus(HttpServletResponse.SC_FORBIDDEN); // 403
+                json.addProperty("status", "forbidden");
+                json.addProperty("message", "Cannot modify a delivered sales invoice.");
+                writeResponse(response, json);
+                return;
+            }
+
             SalesInvoiceDTO dto = gson.fromJson(request.getReader(), SalesInvoiceDTO.class);
             dto.setId(idstr);
+
+            Map<String, List<String>> errors = salesInvoiceService.validate(dto, true);
+            if (!errors.isEmpty()) {
+                response.setStatus(HttpServletResponse.SC_BAD_REQUEST); // 400
+                gson.toJson(Collections.singletonMap("errors", errors), response.getWriter());
+                return;
+            }
+
             SalesInvoice unMaskedInvoice = new SalesInvoice(dto);
+
+            List<InvoiceLineItem> duplicates = salesInvoiceService.findDuplicateLineItems(unMaskedInvoice.getInvoice_line_items());
+            if (!duplicates.isEmpty()) {
+                json.addProperty("status", "error");
+                json.addProperty("message", "Duplicate line items found");
+
+                JsonArray duplicateArray = new JsonArray();
+                for (InvoiceLineItem dup : duplicates)
+                    duplicateArray.add(gson.toJsonTree(dup));
+
+                json.add("duplicates", duplicateArray);
+                response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+                writeResponse(response, json);
+                return; // Stop further processing
+            }
+
+            if(unMaskedInvoice.getStatus() == 1){
+                List<InsufficientStockDTO> insufficientList = salesInvoiceService.validateStockAvailability(unMaskedInvoice.getInvoice_line_items());
+
+                if (!insufficientList.isEmpty()) {
+                    response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+                    json.addProperty("status", "error");
+                    json.addProperty("message", "Insufficient stock for one or more products.");
+                    json.add("insufficient_stock", gson.toJsonTree(insufficientList));
+                    writeResponse(response, json);
+                    return;
+                }
+            }
 
             SalesInvoice updatedInvoice = salesInvoiceService.updateInvoice(unMaskedInvoice);
 
@@ -205,8 +299,15 @@ public class SalesInvoiceServlet extends HttpServlet {
 
         JsonObject json = new JsonObject();
         String idStr = pathInfo.substring(1);
-        Integer id = Integer.parseInt(idStr.substring(4));
         try {
+            PrefixValidator validator = new PrefixValidator();
+            String error = validator.validatePrefixedId(idStr , PrefixValidator.EntityType.INVOICE);
+            if (error != null){
+                throw new NumberFormatException(error);
+            }
+
+            Integer id = Integer.parseInt(idStr.substring(4));
+
             if (salesInvoiceService.deleteInvoice(id)) {
                 json.addProperty("status", "success");
                 json.addProperty("message", "Sales Invoice deleted successfully");
@@ -217,6 +318,12 @@ public class SalesInvoiceServlet extends HttpServlet {
                 json.addProperty("message", "Sales Invoice not found for ID: " + idStr);
                 response.setStatus(HttpServletResponse.SC_NOT_FOUND);
             }
+        } catch (NumberFormatException e) {
+            response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+            json.addProperty("status", "error");
+            json.addProperty("message", "Invalid Invoice ID format.");
+            json.addProperty("error", e.getMessage());
+            e.printStackTrace();
         } catch (Exception e) {
             json.addProperty("status", "error");
             json.addProperty("message", "Failed to process Sales Invoice deletion");
